+++ adbyss_psl/src/lib.rs
@@ -691,7 +691,7 @@ impl Domain {
 	/// ```
 	/// use adbyss_psl::Domain;
 	///
-	/// let a = Domain::new("domain.com").unwrap();
+	/// let a = Domain::new(    "domain.com").unwrap();
 	/// let b = Domain::new("www.domain.com").unwrap();
 	///
 	/// // A is equal to itself.
@@ -703,32 +703,21 @@ impl Domain {
 	/// assert!(! Domain::eq(&a, &b));
 	/// ```
 	pub const fn eq(a: &Self, b: &Self) -> bool {
-		let a = a.as_bytes();
-		let b = b.as_bytes();
-
-		if a.len() == b.len() {
-			let mut idx = 0;
-			while idx < a.len() {
-				if a[idx] != b[idx] { return false; }
-				idx += 1;
-			}
-			true
-		}
-		else { false }
+		slices_eq(a.as_bytes(), b.as_bytes())
 	}
 
 	#[must_use]
 	/// # Same TLD?
 	///
-	/// Returns true if two domains share the same TLD.
+	/// Returns `true` if two domains share the same TLD.
 	///
 	/// ## Examples
 	///
 	/// ```
 	/// use adbyss_psl::Domain;
 	///
-	/// let a = Domain::new("domain.com").unwrap();
-	/// let b = Domain::new("store.domain.com").unwrap();
+	/// let a = Domain::new(       "domain.com").unwrap();
+	/// let b = Domain::new( "store.domain.com").unwrap();
 	/// let c = Domain::new("status.domain.com").unwrap();
 	/// assert!(Domain::eq_tld(&a, &b)); // All three do share a TLD!
 	/// assert!(Domain::eq_tld(&b, &c));
@@ -737,16 +726,51 @@ impl Domain {
 	/// assert!(! Domain::eq_tld(&a, &d)); // Different is different.
 	/// ```
 	pub const fn eq_tld(a: &Self, b: &Self) -> bool {
-		let a = a.tld().as_bytes();
-		let b = b.tld().as_bytes();
+		slices_eq(a.tld().as_bytes(), b.tld().as_bytes())
+	}
 
-		if a.len() == b.len() {
-			let mut idx = 0;
-			while idx < a.len() {
-				if a[idx] != b[idx] { return false; }
-				idx += 1;
-			}
-			true
+	#[must_use]
+	/// # Ends With?
+	///
+	/// Returns `true` if `self` is subordinate or equal to `parent`.
+	///
+	/// ## Examples
+	///
+	/// ```
+	/// use adbyss_psl::Domain;
+	///
+	/// let a = Domain::new(    "cpanel.nodaddy.com").unwrap();
+	/// let b = Domain::new("www.cpanel.nodaddy.com").unwrap();
+	/// let c = Domain::new(     "panel.nodaddy.com").unwrap();
+	/// let d = Domain::new(              "addy.com").unwrap();
+	///
+	/// assert!(a.ends_with(&a)); // Equal.
+	/// assert!(b.ends_with(&a)); // Subordinate.
+	///
+	/// // Parents don't "end with" their children.
+	/// assert!(! a.ends_with(&b));
+	///
+	/// // The following would be true for simple strings, but domain
+	/// // component matching is all or nothing.
+	/// assert!(! a.ends_with(&c));
+	/// assert!(! b.ends_with(&c));
+	/// assert!(! a.ends_with(&d));
+	/// assert!(! b.ends_with(&d));
+	/// ```
+	pub const fn ends_with(&self, parent: &Self) -> bool {
+		let a = self.as_bytes();
+		let b = parent.as_bytes();
+
+		if let Some(diff) = a.len().checked_sub(b.len()) {
+			let (before, after) = a.split_at(diff);
+			(
+				before.is_empty() ||
+				(
+					self.root_start >= diff &&
+					before[diff - 1] == b'.'
+				)
+			) &&
+			slices_eq(after, b)
 		}
 		else { false }
 	}
@@ -1099,6 +1123,22 @@ fn sanitize_email_local_slow(good: &str, maybe: &str, mut last: char) -> Option<
 	else { None }
 }
 
+#[inline]
+/// # Constant Slice Equality.
+///
+/// TODO: replace with simple `a == b` once const.
+const fn slices_eq(a: &[u8], b: &[u8]) -> bool {
+	if a.len() == b.len() {
+		let mut idx = 0;
+		while idx < a.len() {
+			if a[idx] != b[idx] { return false; }
+			idx += 1;
+		}
+		true
+	}
+	else { false }
+}
+
 
 
 #[cfg(test)]
@@ -1157,6 +1197,7 @@ mod tests {
 		t_tld_assert("ide.kyoto.jp", None);
 		t_tld_assert("b.ide.kyoto.jp", Some("b.ide.kyoto.jp"));
 		t_tld_assert("a.b.ide.kyoto.jp", Some("b.ide.kyoto.jp"));
+		t_tld_assert("kobe.jp", None);
 		t_tld_assert("c.kobe.jp", None);
 		t_tld_assert("b.c.kobe.jp", Some("b.c.kobe.jp"));
 		t_tld_assert("a.b.c.kobe.jp", Some("b.c.kobe.jp"));